Fix compiler warnings across runtime and tranx86
Group 1 (macro redefinitions): Add #undef before K_FBAR_{INIT,SYNC,ENROLL,RESIGN}
redefinitions in kitable.h.

Group 2 (undefined behaviour shifts): Fix W_* workspace offset macros in
transputer.h to use -(N << WSH) instead of (-N) << WSH; fix ByteSelectMask
similarly; fix one direct shift in etcrtl.c.

Group 3 (type mismatches): Fix atomic_t.value type mismatch in aarch64/atomics.h
att_safe_cas/att_unsafe_cas (unsigned long vs unsigned int). Add #ifndef guards
for TRUE/FALSE in ukcthreads_types.h.

Group 4 (format specifiers): Fix %d/%x -> %ld/%lx with (long)/(unsigned long) casts
for intptr_t/word-typed regconst fields in rtldump.c, regcolour.c, asm386.c,
asmmips.c, asmppc.c, asmsparc.c, arch386.c, archppc.c, archsparc.c, archx64.c.

Group 5 (unused variables): Fix mobproc.c unused variables in wsmap traversal
loops by replacing decode_entry() assignments with (void)decode_entry() where
values are only needed to advance the map pointer. Fix bind3.c pointer-to-int
cast by switching saved_vars array to intptr_t.

Group 6 (unused functions): Add __attribute__((unused)) to
dump_workspace_mapchain and mcache_clearall in mobproc.c.

Group 7 (packed struct alignment): Suppress -Waddress-of-packed-member for
mp_ctrlblk.mapchain accesses in mobproc.c with GCC diagnostic pragmas; the
field is pointer-sized and naturally aligned in practice.

+++ tools/occ21/be/bind3.c
@@ -31,6 +31,7 @@
 #  include <config.h>
 #endif
 #include <stdio.h>
+#include <stdint.h>
 #include "suplib.h"		/* IMPORTED */
 
 #include "includes.h"
@@ -1792,13 +1793,13 @@ PUBLIC void initallocvars (const BOOL needmap)
  */
 PUBLIC void *sp_saveallocvars (const BOOL needmap)
 {
-	int *saved_vars = (int *)newvec (4 * sizeof (int));
+	intptr_t *saved_vars = (intptr_t *)newvec (4 * sizeof (intptr_t));
 
 #if 0
 fprintf (stderr, "bind3: sp_saveallocvars(): needmap = %d\n", (int)needmap);
 #endif
 	saved_vars[0] = need_wsmap;
-	saved_vars[1] = (int)wsmap;
+	saved_vars[1] = (intptr_t)wsmap;
 	saved_vars[2] = wsmap_cur;
 	saved_vars[3] = wsmap_max;
 
@@ -1816,7 +1817,7 @@ fprintf (stderr, "bind3: sp_saveallocvars(): needmap = %d\n", (int)needmap);
  */
 PUBLIC void sp_restoreallocvars (void *saved)
 {
-	int *saved_vars = (int *)saved;
+	intptr_t *saved_vars = (intptr_t *)saved;
 
 #if 0
 fprintf (stderr, "bind3: sp_restoreallocvars(): current need_wsmap = %d\n", need_wsmap);
@@ -1826,7 +1827,7 @@ fprintf (stderr, "bind3: sp_restoreallocvars(): current need_wsmap = %d\n", need
 	wsmap_cur = saved_vars[2];
 	wsmap_max = saved_vars[3];
 
-	freevec (saved, 4 * sizeof (int));
+	freevec (saved, 4 * sizeof (intptr_t));
 	return;
 }
 /*}}}*/
